modify

+++ CMPageTitleView/CMPageTitleView/Class/CMPageTitleContentView.m
@@ -62,7 +62,7 @@
     
     _cm_selectedIndex = cm_selectedIndex;
     [self selectCellIndex:cm_selectedIndex];
-
+    
     
 }
 
@@ -70,18 +70,18 @@
 - (UIView *)underLine {
     
     if (!_underLine) {
-//        UIView *underLine = [UIView new];
-//        underLine.backgroundColor = self.config.cm_underlineColor;
-//        underLine.layer.cornerRadius = self.config.cm_underlineBorder ? self.config.cm_underlineHeight * 0.5 : 0;
-//        underLine.layer.masksToBounds = YES;
-//        CGFloat underLineWidth = self.config.cm_underlineWidth ?: [self.titleLabels[self.config.cm_selectedIndex] cm_width] * self.config.cm_underlineWidthScale;
-//        underLine.cm_height = self.config.cm_underlineHeight;
-//        underLine.cm_bottom = self.config.cm_titleHeight;
-//        underLine.cm_width  = underLineWidth;
-//        underLine.cm_centerX = [self.titleLabels[self.config.cm_selectedIndex] cm_centerX];
-//        _underLine = underLine;
+//                UIView *underLine = [UIView new];
+//                underLine.backgroundColor = self.config.cm_underlineColor;
+//                underLine.layer.cornerRadius = self.config.cm_underlineBorder ? self.config.cm_underlineHeight * 0.5 : 0;
+//                underLine.layer.masksToBounds = YES;
+//                CGFloat underLineWidth = self.config.cm_underlineWidth ?: [self.titleLabels[self.config.cm_selectedIndex] cm_width] * self.config.cm_underlineWidthScale;
+//                underLine.cm_height = self.config.cm_underlineHeight;
+//                underLine.cm_bottom = self.config.cm_titleHeight;
+//                underLine.cm_width  = underLineWidth;
+//                underLine.cm_centerX = [self.titleLabels[self.config.cm_selectedIndex] cm_centerX];
+//                _underLine = underLine;
         [self addSubview:_underLine];
-
+        
     }
     
     return _underLine ;
@@ -93,20 +93,20 @@
     
     if (!_titleCover) {
         
-//        UIView *titleCover = [UIView new];
-//        titleCover.backgroundColor = self.config.cm_coverColor ?: [UIColor colorWithWhite:0 alpha:0.3];
-//        UILabel *label = self.titleLabels[self.config.cm_selectedIndex];
-//        CGFloat width = [self.config.cm_titleWidths[self.config.cm_selectedIndex] floatValue];
-//        CGFloat coverW = self.config.cm_coverWidth ? : width + 2 * self.config.cm_coverHorizontalMargin;
-//        CGFloat coverH = label.font.pointSize + 2 * self.config.cm_coverVerticalMargin;
-//        titleCover.layer.cornerRadius = self.config.cm_coverRadius ?: coverH * 0.5;
-//        titleCover.cm_height = coverH;
-//        titleCover.cm_width = coverW;
-//        titleCover.cm_centerX = label.cm_centerX;
-//        titleCover.cm_centerY = self.config.cm_titleHeight * 0.5;
-//        _titleCover = titleCover;
-//        [self insertSubview:_titleCover atIndex:0];
-
+        //        UIView *titleCover = [UIView new];
+        //        titleCover.backgroundColor = self.config.cm_coverColor ?: [UIColor colorWithWhite:0 alpha:0.3];
+        //        UILabel *label = self.titleLabels[self.config.cm_selectedIndex];
+        //        CGFloat width = [self.config.cm_titleWidths[self.config.cm_selectedIndex] floatValue];
+        //        CGFloat coverW = self.config.cm_coverWidth ? : width + 2 * self.config.cm_coverHorizontalMargin;
+        //        CGFloat coverH = label.font.pointSize + 2 * self.config.cm_coverVerticalMargin;
+        //        titleCover.layer.cornerRadius = self.config.cm_coverRadius ?: coverH * 0.5;
+        //        titleCover.cm_height = coverH;
+        //        titleCover.cm_width = coverW;
+        //        titleCover.cm_centerX = label.cm_centerX;
+        //        titleCover.cm_centerY = self.config.cm_titleHeight * 0.5;
+        //        _titleCover = titleCover;
+        //        [self insertSubview:_titleCover atIndex:0];
+        
     }
     return _titleCover ;
 }
@@ -152,15 +152,15 @@
 }
 
 - (void)selectCellIndex:(NSInteger)index{
-  _lastOffsetX = index * self.cm_width;
+    _lastOffsetX = index * self.cm_width;
     _cm_selectedIndex = index;
-//    self.config.cm_defaultIndex = index;
+    //    self.config.cm_defaultIndex = index;
     NSIndexPath *indexPath = [NSIndexPath indexPathForItem:index inSection:0];
     
     [self.collectionView selectItemAtIndexPath:indexPath animated:NO scrollPosition:UICollectionViewScrollPositionNone];
-  
+    
     [self modifyCenterWithIndexPath:indexPath];
- 
+    
     
 }
 /**
@@ -168,7 +168,7 @@
  */
 - (void)modifyCenterWithIndexPath:(NSIndexPath *)indexPath{
     self.config.cm_defaultIndex = indexPath.item;
-
+    
     UICollectionViewCell *cell = [self.collectionView cellForItemAtIndexPath:indexPath];
     
     if (self.collectionView.contentSize.width <= self.cm_width) return;
@@ -184,7 +184,7 @@
     
     offsetX = offsetX > maxOffsetX ? maxOffsetX : offsetX;
     
-//    [self.collectionView ffsetContentOffset:CGPointMake(offsetX, 0) animated:YES];
+    //    [self.collectionView ffsetContentOffset:CGPointMake(offsetX, 0) animated:YES];
     [self.collectionView scrollToItemAtIndexPath:indexPath atScrollPosition:UICollectionViewScrollPositionCenteredHorizontally animated:YES];
     
 }
@@ -193,11 +193,11 @@
 #pragma mark --- UICollectionViewDelegate
 
 - (void)scrollViewDidEndDecelerating:(UIScrollView *)scrollView {
-
-//    NSArray *array =  [self.collectionView indexPathsForVisibleItems];
-//    NSInteger idx = ceil(array.count / 2) ;
-//    [self.collectionView scrollToItemAtIndexPath:array[idx] atScrollPosition:UICollectionViewScrollPositionCenteredHorizontally animated:YES];
-
+    
+    //    NSArray *array =  [self.collectionView indexPathsForVisibleItems];
+    //    NSInteger idx = ceil(array.count / 2) ;
+    //    [self.collectionView scrollToItemAtIndexPath:array[idx] atScrollPosition:UICollectionViewScrollPositionCenteredHorizontally animated:YES];
+    
 }
 
 
@@ -206,10 +206,10 @@
 
 - (void)collectionView:(UICollectionView *)collectionView didSelectItemAtIndexPath:(NSIndexPath *)indexPath {
     _isClickTitle = YES;
-
+    
     [self modifyCenterWithIndexPath:indexPath];
     CMPageTitleCell *cell = (CMPageTitleCell*)[collectionView cellForItemAtIndexPath:indexPath];
-    NSLog(@"%@", [NSString stringWithFormat:@"选中 -- %02ld -- %@",indexPath.item,cell.titleLabel.text]);
+    NSLog(@"%@", [NSString stringWithFormat:@"选中 -- %02ld -- %@",(long)indexPath.item,cell.titleLabel.text]);
     
     if (self.cm_delegate) {
         
@@ -224,7 +224,7 @@
 
 - (void)collectionView:(UICollectionView *)collectionView didDeselectItemAtIndexPath:(NSIndexPath *)indexPath {
     CMPageTitleCell *cell = (CMPageTitleCell*)[collectionView cellForItemAtIndexPath:indexPath];
-       NSLog(@"%@", [NSString stringWithFormat:@"反选 -- %02ld -- %@",indexPath.item,cell.titleLabel.text]);
+    NSLog(@"%@", [NSString stringWithFormat:@"反选 -- %02ld -- %@",(long)indexPath.item,cell.titleLabel.text]);
     
 }
 
@@ -247,19 +247,19 @@
     cell.titleLabel.font = self.config.cm_font;
     
     [cell cm_cellSetSelectedCompletion:^(BOOL selected) {
-       
+        
         if (self.config.cm_switchMode & CMPageTitleSwitchMode_Scale){
-               
-               CGFloat scale = 1/ self.config.cm_scale ;
-               UIFont *font = [UIFont systemFontOfSize:self.config.cm_font.pointSize*self.config.cm_scale];
-               cell.titleLabel.font = font;
-               cell.transform = CGAffineTransformMakeScale(scale, scale);
-               
-               if (cell.isSelected) {
-                   cell.transform = CGAffineTransformIdentity;
-               }
-               
-           }
+            
+            CGFloat scale = 1/ self.config.cm_scale ;
+            UIFont *font = [UIFont systemFontOfSize:self.config.cm_font.pointSize*self.config.cm_scale];
+            cell.titleLabel.font = font;
+            cell.transform = CGAffineTransformMakeScale(scale, scale);
+            
+            if (cell.isSelected) {
+                cell.transform = CGAffineTransformIdentity;
+            }
+            
+        }
         
     }];
     
@@ -295,9 +295,9 @@
 
 - (void)cm_pageTitleContentViewAdjustPosition:(UIScrollView *)scrollView {
     NSInteger centerIndex = floorf(scrollView.contentOffset.x / self.cm_width);
-
+    
     [self modifyCenterWithIndexPath:[NSIndexPath indexPathForItem:centerIndex inSection:0]];
-
+    
 }
 
 
@@ -310,77 +310,78 @@
 
 
 - (void)modifyColorWithScrollProgress:(CGFloat)progress FromIndex:(NSUInteger)fromIndex ToIndex:(NSUInteger)toIndex {
-
+    
     if (_isClickTitle || toIndex >= [self.collectionView numberOfItemsInSection:0]) return;
-          
-          CMPageTitleCell *toCell = (CMPageTitleCell *)[self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:toIndex inSection:0]];
-          CMPageTitleCell *fromCell = (CMPageTitleCell *) [self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:fromIndex inSection:0]];
-          CMDisplayTitleLabel *toLabel = toCell.titleLabel;
-          CMDisplayTitleLabel *fromLabel = fromCell.titleLabel;
-
-          CGFloat toScale = fabs(progress);
-          
-          if (self.config.cm_gradientStyle == CMTitleColorGradientStyle_RGB) {
-              
-              NSArray *endRGBA = CMColorGetRGBA(self.config.cm_selectedColor);
-              NSArray *startRGBA = CMColorGetRGBA(self.config.cm_normalColor);
-              
-              CGFloat deltaR = [endRGBA[0] floatValue] - [startRGBA[0] floatValue];
-              CGFloat deltaG = [endRGBA[1] floatValue] - [startRGBA[1] floatValue];
-              CGFloat deltaB = [endRGBA[2] floatValue] - [startRGBA[2] floatValue];
-              CGFloat deltaA = [endRGBA[3] floatValue] - [startRGBA[3] floatValue];
-              
-              
-              UIColor *toColor = [UIColor colorWithRed:[startRGBA[0] floatValue] + toScale *deltaR green:[startRGBA[1] floatValue] + toScale *deltaG blue:[startRGBA[2] floatValue] + toScale *deltaB alpha:[startRGBA[3] floatValue] + toScale *deltaA];
-              
-              UIColor *fromColor = [UIColor colorWithRed:[endRGBA[0] floatValue] - toScale * deltaR green:[endRGBA[1] floatValue] - toScale *deltaG blue:[endRGBA[2] floatValue] + toScale *deltaB alpha:[endRGBA[3] floatValue] - toScale *deltaA];
-
-              toLabel.textColor = toColor;
-              fromLabel.textColor = fromColor;
-              
-          }
-          
-          // 填充渐变
-          if (self.config.cm_gradientStyle == CMTitleColorGradientStyle_Fill) {
-              
-              toLabel.textColor    = progress > 0 ? self.config.cm_normalColor : self.config.cm_selectedColor;
-              toLabel.cm_fillColor = progress > 0 ? self.config.cm_selectedColor : self.config.cm_normalColor;
-              toLabel.cm_progress  = progress > 0 ? toScale : 1-toScale;
-
-              fromLabel.textColor     = progress > 0 ? self.config.cm_selectedColor : self.config.cm_normalColor;
-              fromLabel.cm_fillColor  = progress > 0 ? self.config.cm_normalColor : self.config.cm_selectedColor;
-              fromLabel.cm_progress   = progress > 0 ? toScale : 1-toScale;
-              
-          }
-          
+    
+    CMPageTitleCell *toCell = (CMPageTitleCell *)[self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:toIndex inSection:0]];
+    CMPageTitleCell *fromCell = (CMPageTitleCell *) [self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:fromIndex inSection:0]];
+    CMDisplayTitleLabel *toLabel = toCell.titleLabel;
+    CMDisplayTitleLabel *fromLabel = fromCell.titleLabel;
+    
+    CGFloat toScale = fabs(progress);
+    
+    if (self.config.cm_gradientStyle == CMTitleColorGradientStyle_RGB) {
+        
+        NSArray *endRGBA = CMColorGetRGBA(self.config.cm_selectedColor);
+        NSArray *startRGBA = CMColorGetRGBA(self.config.cm_normalColor);
+        
+        CGFloat deltaR = [endRGBA[0] floatValue] - [startRGBA[0] floatValue];
+        CGFloat deltaG = [endRGBA[1] floatValue] - [startRGBA[1] floatValue];
+        CGFloat deltaB = [endRGBA[2] floatValue] - [startRGBA[2] floatValue];
+        CGFloat deltaA = [endRGBA[3] floatValue] - [startRGBA[3] floatValue];
+        
+        
+        UIColor *toColor = [UIColor colorWithRed:[startRGBA[0] floatValue] + toScale *deltaR green:[startRGBA[1] floatValue] + toScale *deltaG blue:[startRGBA[2] floatValue] + toScale *deltaB alpha:[startRGBA[3] floatValue] + toScale *deltaA];
+        
+        UIColor *fromColor = [UIColor colorWithRed:[endRGBA[0] floatValue] - toScale * deltaR green:[endRGBA[1] floatValue] - toScale *deltaG blue:[endRGBA[2] floatValue] + toScale *deltaB alpha:[endRGBA[3] floatValue] - toScale *deltaA];
+        
+        toLabel.textColor = toColor;
+        fromLabel.textColor = fromColor;
+        
+    }
+    
+    // 填充渐变
+    if (self.config.cm_gradientStyle == CMTitleColorGradientStyle_Fill) {
+        
+        toLabel.textColor    = progress > 0 ? self.config.cm_normalColor : self.config.cm_selectedColor;
+        toLabel.cm_fillColor = progress > 0 ? self.config.cm_selectedColor : self.config.cm_normalColor;
+        toLabel.cm_progress  = progress > 0 ? toScale : 1-toScale;
+        
+        fromLabel.textColor     = progress > 0 ? self.config.cm_selectedColor : self.config.cm_normalColor;
+        fromLabel.cm_fillColor  = progress > 0 ? self.config.cm_normalColor : self.config.cm_selectedColor;
+        fromLabel.cm_progress   = progress > 0 ? toScale : 1-toScale;
+        
+    }
+    
+    
 }
 
 
 
 - (void)modifyScaleWithScrollProgress:(CGFloat)progress FromIndex:(NSUInteger)fromIndex ToIndex:(NSUInteger)toIndex {
     if (!(self.config.cm_switchMode & CMPageTitleSwitchMode_Scale) || _isClickTitle || toIndex >= [self.collectionView numberOfItemsInSection:0] || self.config.cm_switchMode & CMPageTitleSwitchMode_Delay) return;
-   
-        CMPageTitleCell *fromCell = (CMPageTitleCell *)[self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:fromIndex inSection:0]];
-        CMPageTitleCell *toCell = (CMPageTitleCell *) [self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:toIndex inSection:0]];
-
-         CGFloat rightScale = progress;
-
-         CGFloat leftScale = 1 - rightScale;
-
-         CGFloat scaleTransform = 1 - 1/self.config.cm_scale;
-
-         
-         fromCell.transform = CGAffineTransformMakeScale(1 - progress * scaleTransform ,1- progress * scaleTransform );
-         toCell.transform = CGAffineTransformMakeScale(1 - leftScale * scaleTransform , 1 -  leftScale * scaleTransform );
+    
+    CMPageTitleCell *fromCell = (CMPageTitleCell *)[self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:fromIndex inSection:0]];
+    CMPageTitleCell *toCell = (CMPageTitleCell *) [self.collectionView cellForItemAtIndexPath:[NSIndexPath indexPathForItem:toIndex inSection:0]];
+    
+    CGFloat rightScale = progress;
+    
+    CGFloat leftScale = 1 - rightScale;
+    
+    CGFloat scaleTransform = 1 - 1/self.config.cm_scale;
+    
+    
+    fromCell.transform = CGAffineTransformMakeScale(1 - progress * scaleTransform ,1- progress * scaleTransform );
+    toCell.transform = CGAffineTransformMakeScale(1 - leftScale * scaleTransform , 1 -  leftScale * scaleTransform );
 }
 
 - (void)cm_pageContentViewDidScrollProgress:(CGFloat)progress FromIndex:(NSUInteger)fromIndex ToIndex:(NSUInteger)toIndex {
-   
+    
     NSLog(@"滚动时的progress ----- %lf",progress);
     
     [self modifyColorWithScrollProgress:progress FromIndex:fromIndex ToIndex:toIndex];
     [self modifyScaleWithScrollProgress:progress FromIndex:fromIndex ToIndex:toIndex];
-  
+    
     
 }
 
@@ -388,9 +389,9 @@
     
     
     
-//    [self modifyTitleScaleWithScrollProgress:progress LeftIndex:leftIndex RightIndex:rightIndex];
+    //    [self modifyTitleScaleWithScrollProgress:progress LeftIndex:leftIndex RightIndex:rightIndex];
     
-//    [self modifyColorWithScrollProgress:progress LeftIndex:leftIndex RightIndex:rightIndex];
+    //    [self modifyColorWithScrollProgress:progress LeftIndex:leftIndex RightIndex:rightIndex];
     //
     //    [self modifyUnderlineWithScrollProgress:progress LeftIndex:leftIndex RightIndex:rightIndex];
     //
@@ -503,7 +504,7 @@ typedef void(^SelectedBlock)(BOOL selected);
     
     if (!_titleLabel) {
         _titleLabel = [[CMDisplayTitleLabel alloc] initWithFrame:self.bounds];
-//        _titleLabel.font = [UIFont systemFontOfSize:14];
+        //        _titleLabel.font = [UIFont systemFontOfSize:14];
         _titleLabel.textColor = UIColor.blackColor;
         _titleLabel.textAlignment = NSTextAlignmentCenter;
     }
